Handle spaces in redirects
